Module pallet_dkg_metadata::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

force_change_authorities

Forcefully rotate the DKG

force_unjail_keygen

Force removes an authority from keygen jail.

force_unjail_signing

Force removes an authority from signing jail.

set_keygen_threshold

Set the pending keygen threshold for the session following the next session.

set_signature_threshold

Set the pending signature threshold for the session following the next session.

submit_misbehaviour_reports

Submits misbehaviour reports on chain. Signatures of the offending authority are verified against the current or next authorities depending on the type of misbehaviour.

submit_next_public_key

Submits and stores the next public key for the next session into the on-chain storage.

submit_public_key

Submits and stores the active public key for the genesis session into the on-chain storage. This is primarily used to separate the genesis public key submission from non-genesis rounds.

trigger_emergency_keygen

Triggers an Emergency Keygen Protocol.

unjail

Attempts to remove an authority from all possible jails (keygen & signing). This can only be called by the controller of the authority in jail. The origin must map directly to the authority in jail.