Module pallet_dkg_metadata::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AccountToAuthority

Authority account ids scheduled for the next session

Authorities

The current authorities set

AuthorityReputations

Tracks authority reputations

AuthoritySetId

The current authority set id

BestAuthorities

The current best authorities of the active keygen set

CurrentAuthoritiesAccounts

Accounts for the current authorities

CurrentRefreshProposal

The current refresh proposal signed and ready to process

DKGPublicKey

Holds active public key for ongoing session

DKGPublicKeySignature

Signature of the current DKG public key

HistoricalRounds

Tracks current voter set

JailedKeygenAuthorities

Tracks jailed authorities for keygen by mapping to the block number when the authority was last jailed

JailedSigningAuthorities

Tracks jailed authorities for signing by mapping to the block number when the authority was last jailed

KeygenThreshold

The current signature threshold (i.e. the n in t-of-n)

LastSessionRotationBlock

The last BlockNumber at which the session rotation happened

MisbehaviourReports

Tracks misbehaviour reports

NextAuthorities

Authorities set scheduled to be used with the next session

NextAuthoritiesAccounts

Authority account ids scheduled for the next session

NextAuthoritySetId

The next authority set id

NextBestAuthorities

The next best authorities of the active keygen set

NextDKGPublicKey

Holds public key for next session

NextKeygenThreshold

The current signature threshold (i.e. the n in t-of-n)

NextPublicKeySignature

Signature of the DKG public key for the next session

NextSignatureThreshold

The current signature threshold (i.e. the t in t-of-n)

NextUnsignedAt

Defines the block when next unsigned transaction will be accepted.

PendingKeygenThreshold

The pending signature threshold (i.e. the n in t-of-n)

PendingRefreshProposal

The pending refresh proposal waiting for signature

PendingSignatureThreshold

The pending signature threshold (i.e. the t in t-of-n)

PreviousPublicKey

Holds public key for immediate past session

RefreshInProgress

Check if there is a refresh in progress.

RefreshNonce

Nonce value for next refresh proposal

ShouldExecuteNewKeygen

Should we start a new Keygen

ShouldSubmitProposerVote

Should we submit a vote for the new DKG governor if we are a proposer

SignatureThreshold

The current signature threshold (i.e. the t in t-of-n)

UsedSignatures

Public key Signatures for past sessions