Struct pallet_dkg_metadata::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn set_signature_threshold(
origin: OriginFor<T>,
new_threshold: u16
) -> DispatchResultWithPostInfo

Set the pending signature threshold for the session following the next session.

We cannot assume that the next DKG has not already completed keygen. After all, if we are in a new session the next DKG may have already completed. Therefore, when we update the thresholds we are updating a threshold that will become the next threshold after the next session update.

origin - The account origin.

new_threshold - The new signature threshold for the DKG.

pub fn set_keygen_threshold(
origin: OriginFor<T>,
new_threshold: u16
) -> DispatchResultWithPostInfo

Set the pending keygen threshold for the session following the next session.

We cannot assume that the next DKG has not already completed keygen. After all, if we are in a new session the next DKG may have already completed. Therefore, when we update the thresholds we are updating a threshold that will become the next threshold after the next session update.

origin - The account origin.

new_threshold - The new keygen threshold for the DKG.

pub fn set_refresh_delay(
origin: OriginFor<T>,
new_delay: u8
) -> DispatchResultWithPostInfo

Sets the delay when a unsigned RefreshProposal will be added to the unsigned proposal queue.

origin - The account origin.

new_delay - The percentage of elapsed session duration to wait before adding an unsigned refresh proposal to the unsigned proposal queue.

pub fn submit_public_key(
origin: OriginFor<T>,
keys_and_signatures: AggregatedPublicKeys
) -> DispatchResultWithPostInfo

Submits and stores the active public key for the genesis session into the on-chain storage. This is primarily used to separate the genesis public key submission from non-genesis rounds.

Can only be submitted by the current authorities. It is also required that a SignatureThreshold of submissions is reached in order to successfully store the public key on-chain.

origin - The account origin.

keys_and_signatures - The aggregated public keys and signatures for possible current DKG public keys.

pub fn submit_next_public_key(
origin: OriginFor<T>,
keys_and_signatures: AggregatedPublicKeys
) -> DispatchResultWithPostInfo

Submits and stores the next public key for the next session into the on-chain storage.

Can only be submitted by the next authorities. It is also required that a NextSignatureThreshold of submissions is reached in order to successfully store the public key on-chain.

origin - The account origin.

keys_and_signatures - The aggregated public keys and signatures for possible next DKG public keys.

pub fn submit_public_key_signature(
origin: OriginFor<T>,
signature_proposal: RefreshProposalSigned
) -> DispatchResultWithPostInfo

Submits the public key signature for the key refresh/rotation process.

The signature is the signature of the next public key RefreshProposal, signed by the current DKG. It is stored on-chain only if it verifies successfully against the current DKG’s public key. Successful storage of this public key signature also removes the unsigned RefreshProposal from the unsigned queue.

For manual refreshes, after the signature is submitted and stored on-chain, the keys are immediately refreshed and the authority set is immediately rotated and incremented.

origin - The account origin.

signature_proposal - The signed refresh proposal containing the public key signature and nonce.

pub fn submit_misbehaviour_reports(
origin: OriginFor<T>,
reports: AggregatedMisbehaviourReports<T::DKGId>
) -> DispatchResultWithPostInfo

Submits misbehaviour reports on chain. Signatures of the offending authority are verified against the current or next authorities depending on the type of misbehaviour.

Keygen: Verifies against the next authorities, since they are doing keygen.

Signing: Verifies against the current authorities, since they are doing signing.

Verifies the reports against the respective thresholds and if enough reports are met begins to jail and decrease the reputation of the offending authority.

The misbehaviour reputation update is: AUTHORITY_REPUTATION = DECAY_PERCENTAGE * AUTHORITY_REPUTATION

If there are not enough unjailed keygen authorities to perform a keygen after the next session, then we deduct the pending keygen threshold (and pending signing threshold) accordingly.

origin - The account origin.

reports - The aggregated misbehaviour reports containing signatures of an offending authority

pub fn unjail(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Attempts to remove an authority from all possible jails (keygen & signing). This can only be called by the controller of the authority in jail. The origin must map directly to the authority in jail.

The authority’s jail sentence for either keygen or signing must be elapsed for the authority to be removed from the jail.

origin - The account origin.

pub fn force_unjail_keygen(
origin: OriginFor<T>,
authority: T::DKGId
) -> DispatchResultWithPostInfo

Force removes an authority from keygen jail.

Can only be called by the root origin.

origin - The account origin.

authority - The authority to be removed from the keygen jail.

pub fn force_unjail_signing(
origin: OriginFor<T>,
authority: T::DKGId
) -> DispatchResultWithPostInfo

Force removes an authority from signing jail.

Can only be called by the root origin.

origin - The account origin.

authority - The authority to be removed from the signing jail.

pub fn manual_increment_nonce(
origin: OriginFor<T>
) -> DispatchResultWithPostInfo

Manually Update the RefreshNonce (increment it by one).

Can only be called by the root origin.

origin - The account origin. Important: This function is only available for testing purposes.

pub fn manual_refresh(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Manual Trigger DKG Refresh process.

Can only be called by the root origin.

origin - The account that is initiating the refresh process. Important: This function is only available for testing purposes.

pub fn force_change_authorities(
origin: OriginFor<T>
) -> DispatchResultWithPostInfo

Forcefully rotate the DKG

This forces the next authorities into the current authority spot and automatically increments the authority ID. It uses change_authorities to execute the rotation forcefully.

impl<T: Config> Pallet<T>

pub fn deposit_event(event: Event<T>)

impl<T: Config> Pallet<T>

pub fn used_signatures() -> Vec<Vec<u8>>

Public key Signatures for past sessions

impl<T: Config> Pallet<T>

pub fn refresh_nonce() -> u32

Nonce value for next refresh proposal

impl<T: Config> Pallet<T>

pub fn refresh_delay() -> Permill

Session progress required to kickstart refresh process

impl<T: Config> Pallet<T>

pub fn refresh_in_progress() -> bool

Check if there is a refresh in progress.

impl<T: Config> Pallet<T>

pub fn should_manual_refresh() -> bool

Should we manually trigger a DKG refresh process.

impl<T: Config> Pallet<T>

pub fn next_dkg_public_key() -> Option<(AuthoritySetId, Vec<u8>)>

Holds public key for next session

impl<T: Config> Pallet<T>

pub fn next_public_key_signature() -> Option<Vec<u8>>

Signature of the DKG public key for the next session

impl<T: Config> Pallet<T>

pub fn dkg_public_key() -> (AuthoritySetId, Vec<u8>)

Holds active public key for ongoing session

impl<T: Config> Pallet<T>

pub fn public_key_signature() -> Vec<u8>

Signature of the current DKG public key

impl<T: Config> Pallet<T>

pub fn previous_public_key() -> (AuthoritySetId, Vec<u8>)

Holds public key for immediate past session

impl<T: Config> Pallet<T>

pub fn historical_rounds<KArg>(k: KArg) -> RoundMetadata where
KArg: EncodeLike<AuthoritySetId>,

Tracks current proposer set

impl<T: Config> Pallet<T>

pub fn signature_threshold() -> u16

The current signature threshold (i.e. the t in t-of-n)

impl<T: Config> Pallet<T>

pub fn keygen_threshold() -> u16

The current signature threshold (i.e. the n in t-of-n)

impl<T: Config> Pallet<T>

pub fn next_signature_threshold() -> u16

The current signature threshold (i.e. the t in t-of-n)

impl<T: Config> Pallet<T>

pub fn next_keygen_threshold() -> u16

The current signature threshold (i.e. the n in t-of-n)

impl<T: Config> Pallet<T>

pub fn pending_signature_threshold() -> u16

The pending signature threshold (i.e. the t in t-of-n)

impl<T: Config> Pallet<T>

pub fn pending_keygen_threshold() -> u16

The pending signature threshold (i.e. the n in t-of-n)

impl<T: Config> Pallet<T>

pub fn authorities() -> Vec<T::DKGId>

The current authorities set

impl<T: Config> Pallet<T>

pub fn authority_set_id() -> AuthoritySetId

The current authority set id

impl<T: Config> Pallet<T>

pub fn next_authority_set_id() -> AuthoritySetId

The next authority set id

impl<T: Config> Pallet<T>

pub fn next_authorities() -> Vec<T::DKGId>

Authorities set scheduled to be used with the next session

impl<T: Config> Pallet<T>

pub fn current_authorities_accounts() -> Vec<T::AccountId>

Accounts for the current authorities

impl<T: Config> Pallet<T>

pub fn next_authorities_accounts() -> Vec<T::AccountId>

Authority account ids scheduled for the next session

impl<T: Config> Pallet<T>

pub fn account_to_authority<KArg>(k: KArg) -> Option<T::DKGId> where
KArg: EncodeLike<T::AccountId>,

Authority account ids scheduled for the next session

impl<T: Config> Pallet<T>

pub fn misbehaviour_reports<KArg>(
k: KArg
) -> Option<AggregatedMisbehaviourReports<T::DKGId>> where
KArg: EncodeLike<(MisbehaviourType, AuthoritySetId, T::DKGId)>,

Tracks misbehaviour reports

impl<T: Config> Pallet<T>

pub fn authority_reputations<KArg>(k: KArg) -> T::Reputation where
KArg: EncodeLike<T::DKGId>,

Tracks authority reputations

impl<T: Config> Pallet<T>

pub fn jailed_keygen_authorities<KArg>(k: KArg) -> T::BlockNumber where
KArg: EncodeLike<T::DKGId>,

Tracks jailed authorities for keygen by mapping to the block number when the authority was last jailed

impl<T: Config> Pallet<T>

pub fn jailed_signing_authorities<KArg>(k: KArg) -> T::BlockNumber where
KArg: EncodeLike<T::DKGId>,

Tracks jailed authorities for signing by mapping to the block number when the authority was last jailed

impl<T: Config> Pallet<T>

pub fn best_authorities() -> Vec<(u16, T::DKGId)>

The current best authorities of the active keygen set

impl<T: Config> Pallet<T>

pub fn next_best_authorities() -> Vec<(u16, T::DKGId)>

The next best authorities of the active keygen set

impl<T: Config> Pallet<T>

pub fn authority_set() -> AuthoritySet<T::DKGId>

Return the current active DKG authority set.

pub fn next_authority_set() -> AuthoritySet<T::DKGId>

Return the next DKG authority set.

pub fn update_signature_threshold(
new_threshold: u16
) -> DispatchResultWithPostInfo

pub fn update_keygen_threshold(new_threshold: u16) -> DispatchResultWithPostInfo

pub fn decompress_public_key(
compressed: Vec<u8>
) -> Result<Vec<u8>, DispatchError>

pub fn process_public_key_submissions(
aggregated_keys: AggregatedPublicKeys,
authorities: Vec<T::DKGId>
) -> BTreeMap<Vec<u8>, Vec<T::DKGId>>

pub fn process_misbehaviour_reports(
reports: AggregatedMisbehaviourReports<T::DKGId>,
verifying_set: Vec<T::DKGId>
) -> Vec<T::DKGId>

pub fn store_consensus_log(
authority_ids: Vec<T::DKGId>,
next_authority_ids: Vec<T::DKGId>,
active_set_id: AuthoritySetId
)

pub fn should_refresh(now: T::BlockNumber) -> bool

Identifies if a new RefreshProposal should be created at a given block number. This is meant to be called on the on_initialize hook at every block to check if we should begin the refresh proposal signing process.

pub fn insert_historical_refresh(
dkg_pub_key: &(AuthoritySetId, Vec<u8>),
next_pub_key: &(AuthoritySetId, Vec<u8>),
next_pub_key_signature: &[u8]
)

Inserts a successful rotation into the history

Insert historical round metadata consisting of the current round’s public key before rotation, the next round’s public key, and the refresh signature signed by the current key refreshing the next.

pub fn max_extrinsic_delay(_block_number: T::BlockNumber) -> T::BlockNumber

pub fn get_best_authorities_by_reputation(
count: usize,
authorities: &[T::DKGId]
) -> Vec<(u16, T::DKGId)>

pub fn get_best_authorities(
count: usize,
authorities: &[T::DKGId]
) -> Vec<(u16, T::DKGId)>