Trait pallet_dkg_metadata::pallet::Config

pub trait Config: Config + CreateSignedTransaction<Call<Self>> {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type DKGId: Member + Parameter + RuntimeAppPublic + MaybeSerializeDeserialize + AsRef<[u8]> + Into<Public> + From<Public>;
    type KeygenJailSentence: Get<Self::BlockNumber>;
    type SigningJailSentence: Get<Self::BlockNumber>;
    type AuthorityIdOf: Convert<Self::AccountId, Option<Self::DKGId>>;
    type Reputation: Member + Parameter + Default + Encode + Decode + AtLeast32BitUnsigned + Copy;
    type DecayPercentage: Get<Percent>;
    type OffChainAuthId: AppCrypto<Self::Public, Self::Signature>;
    type OnAuthoritySetChangeHandler: OnAuthoritySetChangeHandler<Self::AccountId, AuthoritySetId, Self::DKGId>;
    type OnDKGPublicKeyChangeHandler: OnDKGPublicKeyChangeHandler<AuthoritySetId>;
    type ProposalHandler: ProposalHandlerTrait;
    type NextSessionRotation: EstimateNextSessionRotation<Self::BlockNumber>;
    type RefreshDelay: Get<Permill>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type DKGId: Member + Parameter + RuntimeAppPublic + MaybeSerializeDeserialize + AsRef<[u8]> + Into<Public> + From<Public>

Authority identifier type

type KeygenJailSentence: Get<Self::BlockNumber>

Jail lengths for misbehaviours

type SigningJailSentence: Get<Self::BlockNumber>

type AuthorityIdOf: Convert<Self::AccountId, Option<Self::DKGId>>

Map from controller accounts to their DKG authority identifier.

type Reputation: Member + Parameter + Default + Encode + Decode + AtLeast32BitUnsigned + Copy

The reputation decay percentage

type DecayPercentage: Get<Percent>

The reputation decay percentage

type OffChainAuthId: AppCrypto<Self::Public, Self::Signature>

The identifier type for an offchain worker.

type OnAuthoritySetChangeHandler: OnAuthoritySetChangeHandler<Self::AccountId, AuthoritySetId, Self::DKGId>

Listener for authority set changes

type OnDKGPublicKeyChangeHandler: OnDKGPublicKeyChangeHandler<AuthoritySetId>

type ProposalHandler: ProposalHandlerTrait

type NextSessionRotation: EstimateNextSessionRotation<Self::BlockNumber>

A type that gives allows the pallet access to the session progress

type RefreshDelay: Get<Permill>

Percentage session should have progressed for refresh to begin

type WeightInfo: WeightInfo