Module pallet_dkg_metadata::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

AuthorityIdOf

A Convert implementation that finds the stash of the given controller account, if any.

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Type Definitions

AuthorityReputations

Tracks authority reputations

DKGPublicKey

Holds active public key for ongoing session

DKGPublicKeySignature

Signature of the current DKG public key

HistoricalRounds

Tracks current proposer set

JailedKeygenAuthorities

Tracks jailed authorities for keygen by mapping to the block number when the authority was last jailed

JailedSigningAuthorities

Tracks jailed authorities for signing by mapping to the block number when the authority was last jailed

MisbehaviourReports

Tracks misbehaviour reports

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NextDKGPublicKey

Holds public key for next session

NextPublicKeySignature

Signature of the DKG public key for the next session

PreviousPublicKey

Holds public key for immediate past session

RefreshDelay

Session progress required to kickstart refresh process

RefreshInProgress

Check if there is a refresh in progress.

RefreshNonce

Nonce value for next refresh proposal

ShouldExecuteNewKeygen

Should we start a new Keygen

UsedSignatures

Public key Signatures for past sessions