Trait pallet_dkg_metadata::pallet::Config

pub trait Config: Config + CreateSignedTransaction<Call<Self>> {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type DKGId: Member + Parameter + RuntimeAppPublic + MaybeSerializeDeserialize;
    type OffChainAuthId: AppCrypto<Self::Public, Self::Signature>;
    type OnAuthoritySetChangeHandler: OnAuthoritySetChangeHandler<Self::AccountId, AuthoritySetId, Self::DKGId>;
    type OnDKGPublicKeyChangeHandler: OnDKGPublicKeyChangeHandler<AuthoritySetId>;
    type ProposalHandler: ProposalHandlerTrait;
    type NextSessionRotation: EstimateNextSessionRotation<Self::BlockNumber>;
    type RefreshDelay: Get<Permill>;
    type TimeToRestart: Get<Self::BlockNumber>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type DKGId: Member + Parameter + RuntimeAppPublic + MaybeSerializeDeserialize

Authority identifier type

type OffChainAuthId: AppCrypto<Self::Public, Self::Signature>

The identifier type for an offchain worker.

type OnAuthoritySetChangeHandler: OnAuthoritySetChangeHandler<Self::AccountId, AuthoritySetId, Self::DKGId>

Listener for authority set changes

type OnDKGPublicKeyChangeHandler: OnDKGPublicKeyChangeHandler<AuthoritySetId>

type ProposalHandler: ProposalHandlerTrait

type NextSessionRotation: EstimateNextSessionRotation<Self::BlockNumber>

A type that gives allows the pallet access to the session progress

type RefreshDelay: Get<Permill>

Percentage session should have progressed for refresh to begin

type TimeToRestart: Get<Self::BlockNumber>

Default number of blocks after which dkg keygen will be restarted if it has stalled